tests/requirements.txt
pytest==8.1.1
requests==2.31.0
paho-mqtt==2.0.0
